# apps/device-management/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Smart Home API",
        "description": "REST / Kafka gateway for smart-home platform",
        "version": "2.0"
    },
    "paths": {
...
        "/api/v2.0/automation-scenarios/": {
            "get": {
                "tags": [
                    "AutomationScenarios"
                ],
                "summary": "List Scenarios",
                "operationId": "list_scenarios_api_v2_0_automation_scenarios__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/ScenarioResponse"
                                    },
                                    "type": "array",
                                    "title": "Response List Scenarios Api V2 0 Automation Scenarios  Get"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "AutomationScenarios"
                ],
                "summary": "Create Scenario",
                "operationId": "create_scenario_api_v2_0_automation_scenarios__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ScenarioWithRulesCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ScenarioWithRulesResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v2.0/automation-scenarios/{scenario_id}": {
            "get": {
                "tags": [
                    "AutomationScenarios"
                ],
                "summary": "Get Scenario",
                "operationId": "get_scenario_api_v2_0_automation_scenarios__scenario_id__get",
                "parameters": [
                    {
                        "name": "scenario_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Scenario Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ScenarioWithRulesResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "AutomationScenarios"
                ],
                "summary": "Update Scenario",
                "operationId": "update_scenario_api_v2_0_automation_scenarios__scenario_id__put",
                "parameters": [
                    {
                        "name": "scenario_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Scenario Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ScenarioBase"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ScenarioResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "AutomationScenarios"
                ],
                "summary": "Delete Scenario",
                "operationId": "delete_scenario_api_v2_0_automation_scenarios__scenario_id__delete",
                "parameters": [
                    {
                        "name": "scenario_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Scenario Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ActionType": {
                "type": "string",
                "enum": [
                    "TURN_ON",
                    "TURN_OFF"
                ],
                "title": "ActionType"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "RuleBase": {
                "properties": {
                    "scenario_id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "uuid4"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Scenario Id"
                    },
                    "trigger_type": {
                        "$ref": "#/components/schemas/TriggerType"
                    },
                    "trigger_condition": {
                        "type": "string",
                        "title": "Trigger Condition"
                    },
                    "action_type": {
                        "$ref": "#/components/schemas/ActionType"
                    },
                    "action_target": {
                        "type": "string",
                        "format": "uuid4",
                        "title": "Action Target"
                    }
                },
                "type": "object",
                "required": [
                    "trigger_type",
                    "trigger_condition",
                    "action_type",
                    "action_target"
                ],
                "title": "RuleBase"
            },
            "RuleResponse": {
                "properties": {
                    "scenario_id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "uuid4"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Scenario Id"
                    },
                    "trigger_type": {
                        "$ref": "#/components/schemas/TriggerType"
                    },
                    "trigger_condition": {
                        "type": "string",
                        "title": "Trigger Condition"
                    },
                    "action_type": {
                        "$ref": "#/components/schemas/ActionType"
                    },
                    "action_target": {
                        "type": "string",
                        "format": "uuid4",
                        "title": "Action Target"
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid4",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "trigger_type",
                    "trigger_condition",
                    "action_type",
                    "action_target",
                    "id"
                ],
                "title": "RuleResponse"
            },
            "ScenarioBase": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "user_id": {
                        "type": "string",
                        "format": "uuid4",
                        "title": "User Id"
                    },
                    "enabled": {
                        "type": "boolean",
                        "title": "Enabled",
                        "default": true
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "user_id"
                ],
                "title": "ScenarioBase"
            },
            "ScenarioResponse": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "user_id": {
                        "type": "string",
                        "format": "uuid4",
                        "title": "User Id"
                    },
                    "enabled": {
                        "type": "boolean",
                        "title": "Enabled",
                        "default": true
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid4",
                        "title": "Id"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "user_id",
                    "id",
                    "created_at"
                ],
                "title": "ScenarioResponse"
            },
            "ScenarioWithRulesCreate": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "user_id": {
                        "type": "string",
                        "format": "uuid4",
                        "title": "User Id"
                    },
                    "enabled": {
                        "type": "boolean",
                        "title": "Enabled",
                        "default": true
                    },
                    "rules": {
                        "items": {
                            "$ref": "#/components/schemas/RuleBase"
                        },
                        "type": "array",
                        "title": "Rules"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "user_id",
                    "rules"
                ],
                "title": "ScenarioWithRulesCreate"
            },
            "ScenarioWithRulesResponse": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "user_id": {
                        "type": "string",
                        "format": "uuid4",
                        "title": "User Id"
                    },
                    "enabled": {
                        "type": "boolean",
                        "title": "Enabled",
                        "default": true
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid4",
                        "title": "Id"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "rules": {
                        "items": {
                            "$ref": "#/components/schemas/RuleResponse"
                        },
                        "type": "array",
                        "title": "Rules"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "user_id",
                    "id",
                    "created_at",
                    "rules"
                ],
                "title": "ScenarioWithRulesResponse"
            },
...
            "TriggerType": {
                "type": "string",
                "enum": [
                    "SENSOR",
                    "TIME"
                ],
                "title": "TriggerType"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}